==> SD Tool 2.04/05_1_runOptimizationBYspecies.Rmd <==
---
title: "SDTool (STREAM project): Run Optimization by species"
author: "M.T. Facchini, I. Bitetto (COISPA, Italy)"
date: '`r date()`'
output: word_document
---


```{r compil, eval=F,include=F,cache=F,echo=F,warning=F,message=F,progress=F,verbose=F}
#just a reminder to how compile a docx doc
if(F){
rmarkdown::render("Data.Rmd",c("word_document"))
}
```


#Tools

R, Rstudio and packages.

```{r results='hide', message=FALSE, warning=FALSE}

#chunk option
knitr::opts_chunk$set(cache=FALSE,echo=TRUE, warning=FALSE, 
	message=FALSE, fig.height=4,progress=FALSE,verbose=FALSE,
		include=TRUE,dev='png',autodep=FALSE)

#load packages
library(fishPifct)
library(pander)
library(COSTcore)
library(COSTdbe)
library(COSTeda)
library(data.table) 
library(png)
#library(ggradar)

# SD Tool v.3 and WKBIOPTIM SimTool v.1  were taken into account for the developement of the present tool.
# SD tool has been developed in MARE/2014/19 project and updated in MARE/2016/22 project (STREAM)

# The tool is documented in Deliverable D3.3 "Upgrade the methodological framework and tools for sampling optimization, implement and report case studies" (January 2019)

# Authors of the first version of SD Tool: F. Gontrand, T. Rouyer, N. Billet, 2016
# IFREMER, UMR MARBEC, Avenue Jean Monnet, BP 171, 34203 S` ete, France 

# Authors of the first version of WKBIOPTIM scripts: Nuno Prista (SLU, Sweden) from a subgroup work carried out during a local Workshop on Sampling Design and Optimization (Lysekil, 31 October to 4 November 2016, unpublished)
# SLU Aqua, Institute of Marine Research, Department of Acquatic Resources - Swedish University of Agricultural Sciences

# Authors of this new tool: M.T. Facchini, I. Bitetto, 2017
# COISPA Tecnologia & Ricerca, Via dei Trulli 18/20 - (Bari), Italy 

# In case of use of the tool, the Authors should be cited.
# If you have any comments or suggestions please contact the following e-mail addresses: facchini@coispa.it, bitetto@coispa.it
# SD Tool is believed to be reliable. However, we disclaim any implied warranty or representation 
# about its accuracy, completeness or appropriateness for any particular purpose.



#pander options
panderOptions('table.split.table', 60)
panderOptions('table.style', 'grid')
panderOptions('table.split.cells', 10)
panderOptions('table.alignment.default', 'left')
panderOptions('table.alignment.rownames', 'right')
panderOptions('decimal.mark', ',')
panderOptions('graph.fontsize', '10')


# ATTENTION!!!
# BEFORE STARTING RUN THE CODE FIX THE BUG IN THE COST FUNCTION RUNNING THE FOLLOWING CODE:

fixInNamespace(spdAgreg, pos="package:COSTdbe")

# WHEN THE WINDOWS APPEARS, replace the line 8 with the following code:
###############################################   val <- paste(unlist(BY), collapse = ":-:")

```



# 05 Optimization phase

Run the optimization for a target species in order to find the "optimal" sampling size range analysing the precision of the sampling, i.e. analysing the CV of the length structure of the species. 

This script includes options for a flexible definition of the sampling stratification considering:
a)	different technical strata on the basis of gear (level 4) and/or metier, so grouping strata with similar characteristics; 
b)	different temporal aggregation in order to make flexible the stratification by quarter, depending on fisheries and target species specifications;
c)	different aggregation at spatial level, accounting for the possibility of combining different GSAs (task sharing among GSAs);
d)	data of stocks considered shared among MS, in order to get results on the whole area of the stock (not only by GSA).


##Settings 

```{r settings}
# set the working directory
myWD <- "C:\\Users\\Bitetto Isabella\\OneDrive - Coispa Tecnologia & Ricerca S.C.A.R.L\\MARE22\\STREAM\\FINAL REVISION OF DELIVERABLES\\To upload on GITHUB\\WP3\\Task 3.3\\SD tool 2.04"
setwd(myWD)

CASE_STUDY_PATH <- "C:\\Users\\Bitetto Isabella\\OneDrive - Coispa Tecnologia & Ricerca S.C.A.R.L\\MARE22\\STREAM\\FINAL REVISION OF DELIVERABLES\\To upload on GITHUB\\WP3\\Task 3.3\\SD tool 2.04" 

cl_all <- readRDS(paste(CASE_STUDY_PATH, "/input files/costCL.rds", sep=""))
cs_all <- readRDS(paste(CASE_STUDY_PATH, "/input files/costCS.rds", sep=""))

pastSituations <<- read.csv(paste(CASE_STUDY_PATH, "/04_dataInvestigation/Past situation.csv", sep=""), sep=";")

metier_shortcode <<- read.csv(paste(CASE_STUDY_PATH, "/input files/metier_shortcode.csv", sep=""), sep=";")
table_strat_res <<- read.csv(paste(CASE_STUDY_PATH, "/input files/05_1_stratification_results.csv", sep=""), sep=";")

table_min_accepted_sample_size <<- read.csv(paste(CASE_STUDY_PATH, "/input files/05_1_min_accepted_sample_size.csv", sep=""), sep=";")

 REFERENCE_SPECIES <<- "Parapenaeus longirostris"
 RS_shortcode <<- "DPS"

pastSituations <- pastSituations[pastSituations$Species == REFERENCE_SPECIES, ]

nIter <<- 100


threshold_for_RecyclingRate <<-  0.5

cs_all@ca <- csData()@ca 


  
cs_all@hh$foVal <- "V"

cs_all@sl$landCat <- "HUC"
cs_all@hl$landCat <- "HUC"

cs_all@sl$subSampCat  <- ""
cs_all@hl$subSampCat  <- ""

cs_all@sl$catchCat <- toupper(cs_all@sl$catchCat) 
cs_all@hl$catchCat <- toupper(cs_all@hl$catchCat) 

dir.create(CASE_STUDY_PATH)

for (nr in 1:nrow(cl_all@cl)) {
  if (length(as.character(metier_shortcode$lev4[as.character(metier_shortcode$Fishing_activity_category_level6) == cl_all@cl$foCatEu6[nr]]) ) > 0) {
       cl_all@cl$foCatEu5[nr] =  as.character(metier_shortcode$lev4[as.character(metier_shortcode$Fishing_activity_category_level6) == cl_all@cl$foCatEu6[nr]])
  cl_all@cl$foCatEu6[nr] = as.character(metier_shortcode$short_code[as.character(metier_shortcode$Fishing_activity_category_level6) == cl_all@cl$foCatEu6[nr]]) 
  
} else {
  print(paste("NOT FOUND: ", cl_all@cl$foCatEu6[nr]))
}
} 
  
   for (nr in 1:nrow(cs_all@hh)) {
   if ( length(as.character(metier_shortcode$lev4[as.character(metier_shortcode$Fishing_activity_category_level6) == cs_all@hh$foCatEu6[nr]]) ) >0) {
cs_all@hh$foCatEu5[nr] =  as.character(metier_shortcode$lev4[as.character(metier_shortcode$Fishing_activity_category_level6) == cs_all@hh$foCatEu6[nr]])
  cs_all@hh$foCatEu6[nr] =   as.character(metier_shortcode$short_code[as.character(metier_shortcode$Fishing_activity_category_level6) == cs_all@hh$foCatEu6[nr]])
   } else {
   print(paste("NOT FOUND: ", cs_all@hh$foCatEu6[nr]))    
   } 
   } 

elements <- c("space",	"time",	"technical")
  # elements <- c("space",	"time",	"technical", "commercial_category")

col_ele_var1 <<-  as.character(table_strat_res[1,colnames(table_strat_res) == elements[1] ]) 
col_ele_var2 <<-  as.character(table_strat_res[1,colnames(table_strat_res) == elements[2] ])  
col_ele_var3 <<-  as.character(table_strat_res[1,colnames(table_strat_res) == elements[3] ]) 
# col_ele_var4 <<-  as.character(table_strat_res[1,colnames(table_strat_res) == elements[4] ]) 
 
LengthStr_dataframe <<- data.frame(matrix(nrow=0, ncol=6))    
colnames(LengthStr_dataframe) = c("iteration",  "Var1", "Var2", "Var3", "length", "value")
global_case <<- NULL



```

Input

*CS COST object*

```{r input1}

head(cs_all@tr)
head(cs_all@hh)
head(cs_all@sl)
head(cs_all@hl)
head(cs_all@ca)

```

*Short codes for metiers*

```{r input2}

head(metier_shortcode)

```

*Stratification table*

```{r input3}

table_strat_res

```



ANALYSES & OUTPUT

*Calculation of the CV in the past sampling*

```{r data2 }

gsas <- unique(cs_all@hh$area)

years <- unique(cs_all@tr$year)
years <- years[order(years)]

dummy <- sapply(list.files(pattern="[.]R$", path="SDTfuns/", full.names=TRUE), source)

res_dir <<- paste(CASE_STUDY_PATH, "/05_optimization_",RS_shortcode, sep="")

dir.create(res_dir)


CS_boot <- try(subsetSpp(cs_all, spp == REFERENCE_SPECIES, link = TRUE) , silent=TRUE)
CL_boot <-  subset(cl_all, taxon == REFERENCE_SPECIES, link = TRUE)

 # default setted date : 2999
 CS_boot@tr$year = 2999
 CS_boot@hh$year = 2999
 CS_boot@sl$year = 2999
 CS_boot@hl$year = 2999
 # CL_boot@cl$year = 2999
 
 # sampling
 availSamp = merge(CS_boot@hh, CS_boot@hl)
 
 availLand = CL_boot@cl

 availSamp$SampQuarter <-  quarter(availSamp$date)
   availSamp$SampSemester <- ifelse(quarter(availSamp$date) == 1 | quarter(availSamp$date) == 2, 1, 2)

 availLand$semester <- 1
 availLand$semester[availLand$quarter %in% c(3, 4)] <- 2

 Var1 = ""
 Var2 = ""
 Var3 = ""
 
 Var1_lan = ""
 Var2_lan = ""
 Var3_lan = ""
 
switch(col_ele_var1,
    GSA={ 
    Var1 <- availSamp$area
    Var1_lan <- availLand$area
    },
    Country={
    Var1 <- availSamp$vslFlgCtry
    Var1_lan <- availLand$vslFlgCtry
    },
    ALL = { 
    Var1 <- rep("ALL", length(availSamp$area) )
    Var1_lan <-  rep("ALL", length(availLand$area) )
    },
    stop("Enter something that switches me!")
)


switch(col_ele_var2,
    Q = { 
    Var2 <- availSamp$SampQuarter
    Var2_lan <- availLand$quarter
    },
    S = { 
    Var2 <- availSamp$SampSemester
    Var2_lan <- availLand$semester
    },
    Y = { 
    Var2 <- 1
    Var2_lan <- 1
    },
    stop("Enter something that switches me!")
)


switch(col_ele_var3,    
    lev6 = { 
    Var3 <- availSamp$foCatEu6
    Var3_lan <- availLand$foCatEu6
    },
    lev4 = { 
    Var3 <- availSamp$foCatEu5
    Var3_lan <- availLand$foCatEu5
    },
    NONE = {  
      if (col_ele_var4 == "N") {
         Var3 <- "1"
         Var3_lan <- "1"
      } else {
    print("Sampling by commercial category")
    }
    },
    stop("Enter something that switches me!")
)



# switch(col_ele_var4,    
#     Y = { 
#     Var3 <- availSamp$commCat
#     Var3_lan <- availLand$commCat
#     },
#     N = { print(paste("Sampling by", col_ele_var3) )
#     },
#     stop("Enter something that switches me!")
# )


 availStratSamp = unique(data.frame(Var1 = Var1, Var2 = Var2, Var3 = Var3 ))
 # landings
 availStratLand = unique(data.frame(Var1 = Var1_lan, Var2 =  Var2_lan, Var3 =  Var3_lan))
 # common strata
 tabStrat = merge(availStratSamp, availStratLand)
 
 pander(tabStrat)
 
 ## mean number of individuals by trip codes (samples), for each substratum
 availStrat = Reduce("rbind", lapply(1 : nrow(tabStrat), meanNmeasByTripByStr, tabStrat, CS_boot))
 
 pander(availStrat)
 
 # setting min and max number of trip by stratum
 availStrat$maxSampSize = round(5000/ (availStrat$meanNmeasByTrip))
 availStrat$maxSampSize[availStrat$maxSampSize>1000] = 1000
 availStrat$maxSampSize[availStrat$maxSampSize<100] = 100

 for (nr in 1:nrow(availStrat)) {
resOptCV = try(cvFunStr(nr, strIni(), availStrat, availSamp, CS_boot, CL_boot, 
                    REFERENCE_SPECIES, nIter) , silent=TRUE)
resOptCV$Var1 <-  availStrat$Var1[nr]
resOptCV$Var2 <-  availStrat$Var2[nr]
resOptCV$Var3 <-  availStrat$Var3[nr]

   if (nr == 1) {
      resOptCV_final = resOptCV
   } else {
      resOptCV_final = data.frame(rbind(resOptCV_final, resOptCV))   
   }
   
 }
 
   file_name <- paste( res_dir, "/",RS_shortcode," - CV from bootstrap.csv", sep="") 
  write.table(resOptCV_final,  file_name , sep=";", row.names=F)
  
    file_name <- paste( res_dir, "/",RS_shortcode," - Raised length distribution from bootstrap.csv", sep="") 
write.table( LengthStr_dataframe, file_name , sep=";", row.names=F)
  
 for (i in 1:nrow(availStrat)) {
   
  plot_title <- paste(REFERENCE_SPECIES , " ", availStrat$Var1[i], "-",availStrat$Var2[i], "-",availStrat$Var3[i], " - Bootstrap CVs", sep="")
   file_name <- paste(res_dir, "/",RS_shortcode, " - Bootstrap CVs ", availStrat$Var1[i], "-",availStrat$Var2[i], "-", availStrat$Var3[i] ,".png", sep="") 

   this_str_CV <- resOptCV_final[resOptCV_final$Var1 == availStrat$Var1[i] & resOptCV_final$Var2 == availStrat$Var2[i] & resOptCV_final$Var3 == availStrat$Var3[i], ]
 
    gg <- ggplot(this_str_CV, aes(y=cv, x=sampSize)) + geom_point() + geom_smooth() +
       xlab("# trips") + ylab("coefficient of variation") + ggtitle(plot_title)
    print(gg) 
     ggsave(file_name, gg, width=8, height=6)
 }

  
  for (nr_str in 1:nrow(availStrat)) {
    if (nr_str==1) {
       sampSizeRange <- selectSampRange2(nr_str, availStrat, resOptCV_final)
    } else {
       sampSizeRange <- data.frame(rbind(sampSizeRange, selectSampRange2(nr_str, availStrat, resOptCV_final)))
    }
  }

file_name <- paste( res_dir, "/",RS_shortcode," - Optimized sample size.csv", sep="") 
write.table( sampSizeRange, file_name , sep=";", row.names=F)


for (i in 1:nrow(availStrat)) {

plot_title <- paste(REFERENCE_SPECIES, " ", availStrat$Var1[i], "-",availStrat$Var2[i], "-",availStrat$Var3[i], " - Recycling Rate bootstrap", sep="")
file_name <- paste(res_dir, "/",RS_shortcode, " - Recycling Rate vs no. of trips ", availStrat$Var1[i], "-",availStrat$Var2[i], "-", availStrat$Var3[i] , ".png", sep="") 
   
this_str_CV <- resOptCV_final[as.character(resOptCV_final$Var1) == as.character(availStrat$Var1[i]) & 
                              as.character(resOptCV_final$Var2) == as.character(availStrat$Var2[i]) & 
                              as.character(resOptCV_final$Var3) == as.character(availStrat$Var3[i]), ]

# this_str_CV <- merge(this_str_CV, sample_size_RR, all.x = T)

sampSizeRange_this_str = sampSizeRange[as.character(sampSizeRange$Var1) == as.character(availStrat$Var1[i]) & 
                                         as.character(sampSizeRange$Var2) == as.character(availStrat$Var2[i]) & 
                                         as.character(sampSizeRange$Var3) == as.character(availStrat$Var3[i]), ]
blue_lines <- as.numeric(sampSizeRange_this_str$solutions[!is.na(as.numeric(sampSizeRange_this_str$solutions))])

gg <- ggplot(this_str_CV) + geom_point(mapping=aes(x=sampSize, y=recyclingRate), size = 0.4) +
   geom_smooth(aes(x=sampSize, y=recyclingRate), fill = NA, size = 0.2, col = "black") +
   geom_hline(aes(yintercept = threshold_for_RecyclingRate), col = "red", size = 0.2) +
   theme(axis.text.x = element_text(size=7), axis.text.y = element_text(size=7)) +
   theme(legend.title=element_blank(), legend.position="top") +
   labs(x = "Number of trips", y = "Data Recycling Rate")  + ggtitle(plot_title )

if (length(blue_lines) ==1) {
  gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) 
} else if (length(blue_lines) ==2) {
  gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) +
   geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2)
} else if (length(blue_lines) ==3) {
    gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2)   + 
      geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2) + 
      geom_vline(aes(xintercept = blue_lines[3]), col = "blue", size = 0.2)
} else if (length(blue_lines) ==4) {
      gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) +
   geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2) +
      geom_vline(aes(xintercept = blue_lines[3]), col = "blue", size = 0.2) +
          geom_vline(aes(xintercept = blue_lines[4]), col = "blue", size = 0.2)
}

print(gg) 
ggsave(file_name, gg, width=10, height=8)  

 }


```



*Comparison between optimal sampling size range and past situation*


```{r data3 }
  
 # Here, the code allows to define the optimal range of the sampling size for each substratum. Then, the curves, the past sampling designs, ans the optimal ranges are plotted, and possibly saved.
 
 ## ------------ optimal sample sizes
 # selecting accurate sample size range in each strate

# sampSizeRangeF = data.frame(tabStrat, minSamp = as.numeric(as.character(sampSizeRange[,1])), maxSamp = as.numeric(as.character(sampSizeRange[,2])) )
#  cvFunS4 = merge(resOptCV_final, sampSizeRangeF, by = c("Var1", "Var2", "Var3"))

 # keeping past situations providing cv curves (in availableData)
 
 pastData = merge(pastSituations, availStrat, by = c("Var1", "Var2",  "Var3"))
 
 resOptCV_final$cv[which(is.na(resOptCV_final$cv))] <- 0

 # ATTENTION !!!!! REMOVED ROWS where CV is not available !!!
 # resOptCV_final <- resOptCV_final[resOptCV_final$cv != 0, ] 
  

for (i in 1:nrow(availStrat)) {
  
  plot_title <- paste(REFERENCE_SPECIES, " ",availStrat$Var1[i], "-",availStrat$Var2[i], "-",availStrat$Var3[i],  " - CV vs no. of trips",  sep="")
  file_name <- paste( res_dir, "/",RS_shortcode,  " - CV vs no. of trips ",availStrat$Var1[i], "-",availStrat$Var2[i], "-",availStrat$Var3[i] ,".png", sep="") 

 resOptCV_final_str <- resOptCV_final[as.character(resOptCV_final$Var1) == as.character(availStrat$Var1[i]) & as.character(resOptCV_final$Var2) == as.character(availStrat$Var2[i]) & as.character(resOptCV_final$Var3) == as.character(availStrat$Var3[i]), ]
 pastData_str <-  pastData[as.character(pastData$Var1) == as.character(availStrat$Var1[i]) & as.character(pastData$Var2) == as.character(availStrat$Var2[i]) & as.character(pastData$Var3) == as.character(availStrat$Var3[i]), ]
 
 sampSizeRange_str = sampSizeRange[as.character(sampSizeRange$Var1) == as.character(availStrat$Var1[i]) & as.character(sampSizeRange$Var2) == as.character(availStrat$Var2[i]) & as.character(sampSizeRange$Var3) == as.character(availStrat$Var3[i]), ]
blue_lines <- as.numeric(sampSizeRange_str$solutions[!is.na(as.numeric(sampSizeRange_str$solutions))])

  
 gg <- ggplot(resOptCV_final_str) + 
   geom_point(pastData_str, mapping=aes(x=sampSize, y=cv, col = as.factor(year)), size = 1.8)+
   geom_point(mapping=aes(x=sampSize, y=cv), size = 0.6) +
   geom_smooth(aes(x=sampSize, y=cv), fill = NA, size = 0.2, col = "black") +
   # geom_vline(aes(xintercept = minSamp), col = "blue", size = 0.1)+
   # geom_vline(aes(xintercept = maxSamp), col = "blue", size = 0.1)+
   theme(axis.text.x = element_text(size=7), axis.text.y = element_text(size=7))+
   facet_grid(Var1 ~ Var2, scales='free_x') +
   theme(legend.title=element_blank(), legend.position="top")+
   labs(x = "number of trips", y = "Coefficient of Variation")  + ggtitle(plot_title)
 
 if (length(blue_lines) ==1) {
  gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) 
} else if (length(blue_lines) ==2) {
  gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) +
   geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2)
} else if (length(blue_lines) ==3) {
    gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2)   + 
      geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2) + 
      geom_vline(aes(xintercept = blue_lines[3]), col = "blue", size = 0.2)
} else if (length(blue_lines) ==4) {
      gg <- gg + geom_vline(aes(xintercept = blue_lines[1]), col = "blue", size = 0.2) +
   geom_vline(aes(xintercept = blue_lines[2]), col = "blue", size = 0.2) +
      geom_vline(aes(xintercept = blue_lines[3]), col = "blue", size = 0.2) +
          geom_vline(aes(xintercept = blue_lines[4]), col = "blue", size = 0.2)
}
 
print(gg)
  ggsave(file_name, gg, width=10, height=8)  

} 
 


```
